mcp-server-sdlc-artifacts/requirements.txt
mcp
requests
python-dotenv
pydantic